feat: add caesar and small refactoring
