+++ index.js
@@ -1,5 +1,6 @@
 import * as minimist from 'minimist';
 import { getOptions } from './src/getOptions';
-import { inputStream } from './src/getOptions';
-import { transformStream } from './src/getOptions';
-import { outputStream } from './src/getOptions';
+import { inputStream } from './src/stream/inputStream';
+import { transformStream } from './src/transformStream';
+import { outputStream } from './src/outputStream';
+import { pipeline } from 'stream';
