--- src/coder.js
@@ -1,3 +1,0 @@
-export const coder = ({ mode, shift, text }) => {
-  
-}
