+++ src/stream/inputStream.js
