+++ src/stream/outputStream.js
